+++ CHANGELOG.md
@@ -24,6 +24,7 @@
     - now allows multiple (dynamic) legends for one layer at the same time. Applied to eq-catalogue.
     - now has circle-legend
     - fix: race condition in making post-eq-damage mapable.
+    - fix: wrong style and legend for chile-damage.
 - Compare frontend
     - all code now in modules
     - map allows per-step data-converters which are semi-automatically injected